Refactored interface menus into separate modules
